+++ src/json.erl
@@ -17,7 +17,7 @@
 -export([parse/1, parse/2, serialize/1, serialize/2,
          default_serializers/0]).
 
--export_type([value/0, key/0,
+-export_type([value/0, array/0, object/0, key/0,
               error/0, error_reason/0,
               position/0,
               parsing_options/0, duplicate_key_handling/0,
@@ -25,12 +25,15 @@
               serialization_fun/0, serializers/0]).
 
 -type value() :: null
-               | true | false
+               | boolean()
                | number()
                | binary()
-               | [value()]
-               | #{key() := value()}
+               | array()
+               | object()
                | {atom(), term()}.
+
+-type array() :: [value()].
+-type object() :: #{key() := value()}.
 
 -type key() :: binary() | string() | atom().